fix: 썸네일 엘리먼트가 기존 네이밍 규칙을 유지하도록 이름 변경
